Move Lobby main method to lobby sub-project
