+++ game-core/src/main/java/games/strategy/engine/config/lobby/LobbyPropertyReader.java
@@ -11,7 +11,7 @@ import games.strategy.engine.config.PropertyReader;
  * Reads property values from the lobby configuration file.
  */
 public final class LobbyPropertyReader {
-  private static final String LOBBY_PROPERTIES_FILE = "config/lobby/lobby.properties";
+  private static final String LOBBY_PROPERTIES_FILE = "config/lobby.properties";
 
   private final PropertyReader propertyReader;
