+++ lobby/build.gradle
@@ -22,7 +22,7 @@ apply plugin: 'org.junit.platform.gradle.plugin'
 
 group = 'triplea'
 description = 'TripleA Lobby'
-mainClassName = 'games.strategy.engine.framework.GameRunner'
+mainClassName = 'games.strategy.engine.lobby.server.LobbyRunner'
 
 ext {
     artifactsDir = file("$buildDir/artifacts")
